+++ src/main/kotlin/com/loadingbyte/cinecred/ui/demo/ScreencastDemo.kt
@@ -450,8 +450,9 @@ private class Screencast(
     private val height = desktop.height + captionHeight
 
     private val videoWriter = VideoWriter(
-        file, Resolution(width, height), FPS(fps, 1), avcodec.AV_CODEC_ID_H264, avutil.AV_PIX_FMT_YUV420P, emptyMap(),
-        codecOptions = mapOf("crf" to "17")
+        file, Resolution(width, height), FPS(fps, 1), avcodec.AV_CODEC_ID_H264, avutil.AV_PIX_FMT_YUV420P,
+        VideoWriter.Range.LIMITED, VideoWriter.TransferCharacteristic.BT709, VideoWriter.YCbCrCoefficients.BT709,
+        emptyMap(), codecOptions = mapOf("crf" to "17")
     )
 
     private var caption = mutableListOf<TextLayout>()